+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Unreleased
 
+* Implement better support for Windows shell integrations.
+    - Add `zng::env::About::windows_aumid`.
+    - View-process now explicitly associates the AUMID with the GUI process.
+
 * Fix `command_property!` generated *can_cmd* not implementing for widget when requested.
 * Add `zng-wgt-wizard` and `zng::wizard` with the `Wizard!` widget.
 * Add `zng-ext-setup` crate for creating (un)installers.